refactor(view-manager): extract BaseViewManager template-method class
WebContentsViewManager was ~1300 lines mixing the shared UI/state-machine
layer with all the Electron WebContentsView I/O. Split it so an
alternative implementation can subclass the base without duplicating
fragile coordination logic.

BaseViewManager (new) owns:
- mode + active-workspace + loading bookkeeping
- attach-before-detach sequencing on workspace switches
- z-order re-raise rules for dialog/shortcut/hover
- focus routing (mode + attachment state)
- bounds math via the shared helpers
- workspace registry + event subscriptions + reentrancy guard
- destroy/teardown orchestration
- reloadAllViews default (iterate + delegate to subclass)

WebContentsViewManager now contains only the Electron-specific bits it
genuinely needs: session/partition setup, security webPreferences,
event-handler wiring (will-navigate, did-fail-load, render-process-gone,
did-finish-load, unresponsive/responsive), exponential-backoff retry,
the reload watchdog with the recreate fallback, and the Windows
DirectComposition re-composite workaround. Everything else flows
through ~15 protected abstract primitives.

No behavior change: all 3755 tests pass unchanged.
